Made editor pay attention to mouse buttons, and allowed for scrolling Added a settings menu, currently blank but can be populated soon Added loading of resolution and fullscreen state from xml file Added death functions to allow sections to spawn other sections on death, contribute to a score etc Added a Geometry Wars knock-off Added XMLCore to facilitate GeometryWars.luaChallenge Removed some cruft from JuggernautFight.luaChallenge Removed some compiler warnings

+++ OGLTest_lib/Section.cpp
@@ -345,7 +345,7 @@ bool Section::CheckCollisions(const Vector3f _lineP1, const Vector3f _lineP2, Ve
 
 	Vector3f collision_point;
 	Vector3f* first_point = &transformed_outline_verts_[0];
-	int num_points = outline_.GetOutlineVerts()->size();
+	int num_points = static_cast<int>(outline_.GetOutlineVerts()->size());
 
 	if(Collisions2f::LineInPolygon(_lineP1, _lineP2, first_point, num_points, collision_point))
 	{
